common-ize Chrysler checksum functions
